organize auth files
