+++ src/authentication/custom_strategies/basic_auth.js
